CRITICAL FIX: Cast database port to int for PHP 8.4 compatibility
Root Cause:
PHP 8.4 has stricter type checking for mysqli::real_connect().
The port parameter (argument #5) MUST be int|null, but env() was
returning it as a string '3306', causing a TypeError.

Error:
mysqli::real_connect(): Argument #5 ($port) must be of type ?int, string given

Fix:
Cast env('database.default.port') to (int) in Database.php constructor.

This was preventing ANY database connection, which blocked login completely.
The session code was correct - the error was happening before login could
even check credentials.

Impact:
- Login now works (can connect to database)
- All database operations restored
- Session persistence can now be tested properly

PHP Version: 8.4.14 (requires strict type compliance)

+++ app/Config/Database.php
@@ -159,7 +159,7 @@ class Database extends Config
         $this->default['password'] = env('database.default.password', '');
         $this->default['database'] = env('database.default.database', 'ponto_eletronico');
         $this->default['DBDriver'] = env('database.default.DBDriver', 'MySQLi');  // Default to MySQL for production
-        $this->default['port']     = env('database.default.port', 3306);  // MySQL port
+        $this->default['port']     = (int) env('database.default.port', 3306);  // CRITICAL: Cast to int for PHP 8.4 compatibility
         $this->default['charset']  = env('database.default.charset', 'utf8mb4');
         $this->default['DBCollat'] = env('database.default.DBCollat', 'utf8mb4_general_ci');
